Make Hand class into Handable module
Using a mixin rather than inheritance feels better because there
was never actually a situation in which you would want a Hand
object. Instead, you would always want an instance of one of it's
subclasses. As a mixin, it is explicit about the fact that it
should not be instantiated.